requirements.txt
pandas
requests
qdrant-client
gradio
numpy
torch
torchvision
facenet-pytorch
Pillow
scikit-learn
tqdm
matplotlib
pyyaml